// src/app/model/av.ts
export type Av = {
    power?: number;
    menu?: number;
    cancel?: number;
    info?: number;

    volumeUp?: number;
    volumeDown?: number;
    channelUp?: number;
    channelDown?: number;

    fastRewind?: number;
    fastForward?: number;
    skipPrevious?: number;
    play?: number;
    skipNext?: number;
    stopRecord?: number;
    pause?: number;
    stop?: number;

    key1?: number;
    key2?: number;
    key3?: number;
    key4?: number;
    key5?: number;
    key6?: number;
    key7?: number;
    key8?: number;
    key9?: number;
    key0?: number;
    asterisk?: number;
    hash?: number;

    up?: number;
    down?: number;
    left?: number;
    right?: number;
    enter?: number;

    mute?: number;
    red?: number;
    yellow?: number;
    green?: number;
    cyan?: number;
};